# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo List API",
        "description": "This is the API documentation for the Todo List application. It provides endpoints to manage todo tasks, including creating, retrieving, updating, and deleting tasks.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/auth": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate user and generate token",
                "operationId": "5b52d28075e8bd24dc4cf99e9ba0403f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful authentication",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/todo-tasks": {
            "get": {
                "tags": [
                    "TodoTasks"
                ],
                "summary": "Get a list of todo tasks for the authenticated user",
                "operationId": "333436b5a8b4c44d18b5dd5ae6b75cec",
                "parameters": [
                    {
                        "name": "by_status",
                        "in": "query",
                        "description": "Filter tasks by status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of todo tasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TodoTask"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "TodoTasks"
                ],
                "summary": "Create a new todo task",
                "operationId": "1f4187c577c07c7121a4357468d4277e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TodoTask"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Todo task created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TodoTask"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/todo-tasks/{id}": {
            "get": {
                "tags": [
                    "TodoTasks"
                ],
                "summary": "Get a specific todo task by ID",
                "operationId": "3af1d94a4fbce32ae0333ede66db943c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the todo task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todo task details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TodoTask"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Todo task not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "TodoTasks"
                ],
                "summary": "Update the status of a todo task",
                "operationId": "549db1ccc259f779544894853d40a286",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the todo task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "The new status for the todo task",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Todo task status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TodoTask"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid status provided"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "TodoTasks"
                ],
                "summary": "Delete a todo task",
                "operationId": "a5a1bfdf734ceb7f5d8b1ce074142e37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the todo task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Todo task deleted successfully"
                    },
                    "404": {
                        "description": "Todo task not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Register a new user",
                "operationId": "6409506927eb5a335ed6869a59fde03e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "TodoTask": {
                "title": "TodoTask",
                "required": [
                    "title",
                    "description",
                    "user_id",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "status": {
                        "description": "Todo Task Status",
                        "type": "string",
                        "enum": [
                            "Pendente",
                            "Em andamento",
                            "Concluída"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "TodoTasks",
            "description": "TodoTasks"
        },
        {
            "name": "Register",
            "description": "Register"
        }
    ]
}